Shut up clion code analysis

+++ fahbench/web/request_handler.cpp
@@ -72,7 +72,7 @@ void request_handler::handle_file_request(const std::string & request_path, repl
     rep.status = reply::ok;
     char buf[512];
     while (is.read(buf, sizeof(buf)).gcount() > 0)
-        rep.content.append(buf, is.gcount());
+        rep.content.append(buf, (unsigned long) is.gcount());
     rep.headers.resize(2);
     rep.headers[0].name = "Content-Length";
     rep.headers[0].value = std::to_string(rep.content.size());
